added company and update account endpoints, update portfolio, get countries, get states

+++ sources/dronesapp-ws/src/main/java/com/drones4hire/dronesapp/ws/controllers/PortfolioController.java
@@ -2,6 +2,7 @@ package com.drones4hire.dronesapp.ws.controllers;
 
 import com.drones4hire.dronesapp.models.db.portfolio.PortfolioItem;
 import com.drones4hire.dronesapp.models.dto.PortfolioItemDTO;
+import com.drones4hire.dronesapp.services.exceptions.ForbiddenOperationException;
 import com.drones4hire.dronesapp.services.services.PortfolioService;
 import com.drones4hire.dronesapp.ws.swagger.annotations.ResponseStatusDetails;
 import io.swagger.annotations.*;
@@ -12,6 +13,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.ArrayList;
 import java.util.List;
 
 import javax.validation.Valid;
@@ -23,59 +26,89 @@ public class PortfolioController extends AbstractController
 {
 	@Autowired
 	private PortfolioService portfolioService;
-	
+
 	@Autowired
 	private Mapper mapper;
 
 	@ResponseStatusDetails
-	@ApiOperation(value = "Create portfolio item", nickname = "createPortfolioItem", code = 201, httpMethod = "POST", response = PortfolioItem.class)
-	@ApiImplicitParams({@ApiImplicitParam(name = "Authorization", paramType = "header")})
+	@ApiOperation(value = "Create portfolio item", nickname = "createPortfolioItem", code = 201, httpMethod = "POST", response = PortfolioItemDTO.class)
+	@ApiImplicitParams({ @ApiImplicitParam(name = "Authorization", paramType = "header") })
 	@ResponseStatus(HttpStatus.CREATED)
 	@RequestMapping(method = RequestMethod.POST, produces = MediaType.APPLICATION_JSON_VALUE)
-	public @ResponseBody PortfolioItem createPortfolioItem(@Valid @RequestBody PortfolioItemDTO pi)
+	public @ResponseBody PortfolioItemDTO createPortfolioItem(@Valid @RequestBody PortfolioItemDTO pi)
 	{
 		PortfolioItem portfolioItem = mapper.map(pi, PortfolioItem.class);
 		portfolioItem.setUserId(getPrincipal().getId());
-		return portfolioService.createPortfolioItem(portfolioItem);
+		return mapper.map(portfolioService.createPortfolioItem(portfolioItem), PortfolioItemDTO.class);
 	}
 
 	@ResponseStatusDetails
 	@ApiOperation(value = "Get portfolio item by id", nickname = "createPortfolioItem", code = 200, httpMethod = "GET", response = PortfolioItem.class)
-	@ApiImplicitParams({@ApiImplicitParam(name = "Authorization", paramType = "header")})
+	@ApiImplicitParams({ @ApiImplicitParam(name = "Authorization", paramType = "header") })
 	@ResponseStatus(HttpStatus.OK)
 	@RequestMapping(value = "{id}", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
-	public @ResponseBody PortfolioItem getPortfolioItemById(@ApiParam(value = "Id of the portfolio", required = true) @PathVariable(value="id") long id)
+	public @ResponseBody PortfolioItem getPortfolioItemById(
+			@ApiParam(value = "Id of the portfolio", required = true) @PathVariable(value = "id") long id)
 	{
 		return portfolioService.getPortfolioItemById(id);
 	}
 
 	@ResponseStatusDetails
 	@ApiOperation(value = "Get user portfolio items", nickname = "getUserPortfolioItems", code = 200, httpMethod = "GET", response = List.class)
-	@ApiImplicitParams({@ApiImplicitParam(name = "Authorization", paramType = "header")})
+	@ApiImplicitParams({ @ApiImplicitParam(name = "Authorization", paramType = "header") })
 	@ResponseStatus(HttpStatus.OK)
 	@RequestMapping(method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
-	public @ResponseBody List<PortfolioItem> getUserPortfolioItems()
+	public @ResponseBody List<PortfolioItemDTO> getUserPortfolioItems()
 	{
-		return portfolioService.getPortfolioItemsByUserId(getPrincipal().getId());
+		List<PortfolioItem> portfolioItems = portfolioService.getPortfolioItemsByUserId(getPrincipal().getId());
+		List<PortfolioItemDTO> portfolioItemDTOs = new ArrayList<>();
+		for (PortfolioItem portfolioItem : portfolioItems)
+		{
+			portfolioItemDTOs.add(mapper.map(portfolioItem, PortfolioItemDTO.class));
+		}
+		return portfolioItemDTOs;
 	}
 
 	@ResponseStatusDetails
-	@ApiOperation(value = "Update portfolio item", nickname = "updatePortfolioItem", code = 200, httpMethod = "PUT", response = PortfolioItem.class)
-	@ApiImplicitParams({@ApiImplicitParam(name = "Authorization", paramType = "header")})
+	@ApiOperation(value = "Update portfolio item", nickname = "updatePortfolioItem", code = 200, httpMethod = "PUT", response = PortfolioItemDTO.class)
+	@ApiImplicitParams({ @ApiImplicitParam(name = "Authorization", paramType = "header") })
 	@ResponseStatus(HttpStatus.OK)
 	@RequestMapping(method = RequestMethod.PUT, produces = MediaType.APPLICATION_JSON_VALUE)
-	public @ResponseBody PortfolioItem updatePortfolioItem(@RequestBody PortfolioItem portfolioItem)
+	public @ResponseBody PortfolioItemDTO updatePortfolioItem(@RequestBody PortfolioItemDTO pi)
+			throws ForbiddenOperationException
 	{
-		return portfolioService.updatePortfolioItem(portfolioItem);
+		PortfolioItem portfolioItem = portfolioService.getPortfolioItemById(pi.getId());
+		if (portfolioItem.getUserId() == getPrincipal().getId())
+		{
+			portfolioItem.setName(pi.getName());
+			portfolioItem.setTitle(pi.getTitle());
+			portfolioItem.setSummary(pi.getSummary());
+			portfolioItem.setType(pi.getType());
+			portfolioItem.setItemURL(pi.getItemURL());
+			portfolioItem.setServiceCategories(pi.getServiceCategories());
+		} else
+		{
+			throw new ForbiddenOperationException("Invalid user");
+		}
+		return mapper.map(portfolioService.updatePortfolioItem(portfolioItem), PortfolioItemDTO.class);
 	}
 
 	@ResponseStatusDetails
 	@ApiOperation(value = "Delete portfolio item", nickname = "deletePortfolioItem", code = 204, httpMethod = "DELETE")
-	@ApiImplicitParams({@ApiImplicitParam(name = "Authorization", paramType = "header")})
+	@ApiImplicitParams({ @ApiImplicitParam(name = "Authorization", paramType = "header") })
 	@ResponseStatus(HttpStatus.NO_CONTENT)
 	@RequestMapping(value = "{id}", method = RequestMethod.DELETE, produces = MediaType.APPLICATION_JSON_VALUE)
-	public void deletePortfolioItem(@ApiParam(value = "Id of the portfolio", required = true) @PathVariable(value="id") long id)
+	public void deletePortfolioItem(
+			@ApiParam(value = "Id of the portfolio", required = true) @PathVariable(value = "id") long id)
+			throws ForbiddenOperationException
 	{
-		portfolioService.deletePortfolioItem(id);
+		PortfolioItem portfolioItem = portfolioService.getPortfolioItemById(id);
+		if (portfolioItem.getUserId() == getPrincipal().getId())
+		{
+			portfolioService.deletePortfolioItem(id);
+		} else
+		{
+			throw new ForbiddenOperationException("Invalid user");
+		}
 	}
 }
